Completed Day 08
Part 1 was easy and fun to work through. Part 2 was more of a challenge.
I naively thought the same algorithm would work, ran it for over 20
hours, and it still did not complete with an answer.

I had to get a hint about how to tackle part 2. I had no idea you could
solve it with math and LCM, let alone how to implement LCM. So I learned
a bit, but still don't understand how one would figure that out.