Completed Day 10
This challenge was very difficult for me, because I'm not used to any
graph theory, or working with 2d matrices. It took a while to wrap my
head around moving in a 2d space, and then how to dertermine which spot
in the pipe we were on.

Part 2 was extremely difficult, I had no strategy and couldn't not
immediately understand how to go about solving for it. I ended up
looking at tips, and learning about:

  * Pick's Theorem, which can be used to calculate the area of a polygon.
  * Learned that the pipes in the problem were actually just creating a
    polygon that could then be solved for using the area of the polygon.
  * Solve for the area of a polygon by iterating over every point in the
    path and and apply the Shoestring Formula to calculate the area
    which then could be used in Pick's Theorem.